Allow duplicate source packages to be published
Pass the --ignore=updatedsource to reprepro when running includedsc so
that source packages of the same name but different checksums can be
included in multiple distributions.

For Endless, this problem occurs when source packages are generated on
stable branches from git. If the version is the same as master, the
files will be rejected when OBS publishes them. Since the content of the
source packages should be the same, we can simply use the existing files
from master.

[endlessm/eos-shell#4528]