[backend] Add admin feature to purge deleted project
Sometimes you really want to delete a project and the source files that
only it was using. Add a --purge-project feature to bs_admin to allow
fully deleting projects and any source files that aren't referenced by
other projects.

https://github.com/endlessm/eos-shell/issues/4329

FIXME: This duplicates a lot of code from the --remove-old-sources
feature. The code should be refactored so both features can use the same
code.