[backend] Only define buildinfo profile for non-empty value
The default buildconfig profile setting is '' in obs-build. Passing this
through to the buildinfo will eventually cause the build to break
because no argument is supplied with the --profile option.

This was a fix that was only done in the source package patch and wasn't
synced back into 3b46f32081edeaac52fc01c6c721133e782151c0.