[backend] bs_dodup: support | delimited debian URLs
The debian distro and components are currently parsed by taking trailing
parts of the URL path split with /. This means that the distro and
components cannot contain /s, but both are legal characters in debian
repository specifiers. For instance, the debian security repository uses
distro names of <release>/updates as described in
https://www.debian.org/security/#keeping-secure.

To support this, allow a URL syntax where the distro and components are
delimited by |. For example, a URL for the above security repo would be
http://security.debian.org/debian-security/|stable/updates|main.

https://github.com/openSUSE/open-build-service/issues/5217

DBN: This has not been merged upstream so we may have to change format
again when that's decided.

https://phabricator.endlessm.com/T14726