obsworker: Ignore failures from killall
Unlike upstream, debian's obsworker init script runs with -e. This is an
issue on stopping since killall will return 1 if no processes were
found. This is normal if bs_worker --exit did the right thing. Ignore
the exit status of killall so that stop continues to the screen quit
command. Systemd will still regard the service as failed if killall
failed to kill some processes.